Test that the default HSTS header is set when using a proxy header too
The existing `test_hsts_defaults()` test only checked the HSTS header
when using `wsgi.url_scheme` and not a proxy header. In addition it
duplicated part of the `test_https_doesnt_redirect()` test.

As such this moves the default HSTS tests into the redirect tests,
avoiding the duplication and increasing coverage.